requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.4
djangorestframework-stubs==0.4.2
gunicorn==19.9.0
idna==2.8
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
soupsieve==1.9.2
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.3
whitenoise==4.1.2
